Put selector accordion components into separated files
